Project Overview
The web application is an online language school. The online language school contains a catalog of courses. The administrator manages the catalog, processes subscriptions and controls user access. To place a subsctription, a student needs to register using an up-to-date e-mail, after which it becomes available to manage his subscriptions and profile.